+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## 0.1.1.dev0 (Work In Progress)
 
--
+- switched to tracemalloc for general memory tracing
 
 ## 0.1.0 ()